final/model1/requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.27.3